Review Phase 2a: extract KafkaAuth + ConnectorCreator from CLI
Both classes were domain logic living inside connect_migrate.cli.migrate_cli,
making them un-importable from anything that didn't want to drag in the
CLI module. Moved verbatim to
connect_migrate.migration.connector_creator (no behavior change).

cli/migrate_cli.py shrinks from 392 to 215 lines and is now just
argparse + main() orchestration. Drops base64, requests, json (other
than the result-file dump), and the Dict/List/Any/Optional typing
imports that only the moved classes needed.

92/92 unit tests pass.
